--- src/contexts/CountriesContext.ts
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const CountriesContext = createContext(null);
-
-export default CountriesContext;
